# reservas-apiv1/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reservas API",
        "description": "API para sistema de reservas con pagos PayPal y autenticación Sanctum",
        "contact": {
            "email": "soporte@tusistema.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/consultants": {
            "get": {
                "tags": [
                    "Consultores"
                ],
                "summary": "Listar asesores paginados",
                "operationId": "b623671b25df157a58c2ecd74c06af33",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de consultores paginado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Listado obtenido correctamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "nombres": {
                                                                "type": "string",
                                                                "example": "Carlos"
                                                            },
                                                            "apellidos": {
                                                                "type": "string",
                                                                "example": "Ramirez"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "consultor@test.com"
                                                            },
                                                            "telefono": {
                                                                "type": "string",
                                                                "example": "987654321"
                                                            },
                                                            "foto": {
                                                                "type": "string",
                                                                "example": "foto.jpg"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/consultants/{id}/calendar": {
            "get": {
                "tags": [
                    "Consultores"
                ],
                "summary": "Obtener calendario dinámico de un consultor por fecha",
                "operationId": "fda8d9a003ac8ef92713afda49b1d476",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del consultor",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Fecha en formato YYYY-MM-DD",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "2026-03-20"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calendario generado correctamente"
                    },
                    "400": {
                        "description": "Fecha inválida"
                    },
                    "404": {
                        "description": "Consultor no encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/consultant/reservations": {
            "get": {
                "tags": [
                    "Consultores"
                ],
                "summary": "Listar reservas del consultor autenticado",
                "operationId": "f2a6a9ce7d70aee323e4d5e3aa4c709a",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filtrar por fecha YYYY-MM-DD",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "2024-10-23"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de reservas obtenido correctamente"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/consultants/{id}": {
            "get": {
                "tags": [
                    "Consultores"
                ],
                "summary": "Obtener detalle de un consultor",
                "operationId": "fc9b7f212be5753ce098c88c6e433a1e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del consultor",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del consultor obtenido correctamente"
                    },
                    "404": {
                        "description": "Consultor no encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/consultant/calendar": {
            "get": {
                "tags": [
                    "Consultores"
                ],
                "summary": "Obtener calendario del consultor autenticado",
                "operationId": "e3e5c8c8265e2b6f3ac6a0bc7cc6404f",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Fecha en formato YYYY-MM-DD",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "2024-10-23"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calendario obtenido correctamente"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "400": {
                        "description": "Fecha inválida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token en formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}